+++ changelog.md
@@ -2,6 +2,7 @@ Some smaller commits related to minor fixes (specially comment corrections) are 
 
 #### 2020 Jun 04
 * (Network) Maximum snapshot history size can now be configured with a different size for clients and servers. By default 60 and 120, respectively.
+* (Network) Analog input can now be quantized (8-bit precision). It is a project setting that can be enabled (by default it's not used). This creates a new interdependency on the `quantize.gd` script.
 
 #### 2020 Jun 03
 * (Network) Added a "prediction counter" that can be used to re-simulate entities that don't need input data when correction is triggered (`network.snapshot_data.get_prediction_count()`). Tutorial has been updated to show how to use this (topic `Prediction Without Input Data`).